extractor/requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
PyMuPDF==1.26.5
pytesseract==0.3.13
pyzbar==0.1.9
qrcode==8.2
rembg
onnxruntime
